Add analytics service mock
Also renames Logging library's analytics service mock.
