+++ Tests/govuk_ios/govuk_ios_unit_tests/Mocks/Services/MockAnalyticsService.swift
@@ -1,55 +1,29 @@
 import XCTest
 import Logging
 
-final class MockAnalyticsService: Logging.AnalyticsService {
-    var additionalParameters = [String: Any]()
-    private(set) var screensVisited = [String]()
-    private(set) var screenParamsLogged: [String: Any]?
-    private(set) var eventsLogged = [String]()
-    private(set) var eventsParamsLogged: [String: Any]?
-    private(set) var crashesLogged = [NSError]()
-    var hasAcceptedAnalytics: Bool?
-    
-    func trackScreen(_ screen: LoggableScreen, title: String?) {
-        screensVisited.append(screen.name)
+@testable import govuk_ios
+
+class MockAnalyticsService: AnalyticsServiceInterface {
+    private(set) var screensVisited = [Logging.LoggableScreenV2]()
+    private(set) var eventsLogged = [govuk_ios.AppEvent]()
+
+    func configure() {}
+
+    func track(event: govuk_ios.AppEvent) {
+        eventsLogged.append(event)
     }
-    
-    func trackScreen(_ screen: LoggableScreen, parameters: [String: Any] = [:]) {
-        screensVisited.append(screen.name)
-        
-        screenParamsLogged = parameters
+
+    func trackScreen(_ screen: any Logging.LoggableScreenV2) {
+        screensVisited.append(screen)
     }
-    
-    func trackScreen(_ screen: LoggableScreenV2, parameters: [String: Any]) {
-        screensVisited.append(screen.name)
-        
-        var screenParameters = parameters
-        screenParameters["AnalyticsParameterScreenClass"] = screen.type.name
-        screenParameters["AnalyticsParameterScreenName"] = screen.name
-        
-        screenParamsLogged = screenParameters
+
+    var _stubbedAcceptedAnalytics: Bool = true
+    func setAcceptedAnalytics(accepted: Bool) {
+        _stubbedAcceptedAnalytics = accepted
     }
-    
-    
-    func logEvent(_ event: LoggableEvent, parameters: [String: Any]) {
-        eventsLogged.append(event.name)
-        
-        eventsParamsLogged = parameters
-    }
-    
-    func logCrash(_ crash: NSError) {
-        crashesLogged.append(crash)
-    }
-    
-    func logCrash(_ crash: Error) {
-        crashesLogged.append(crash as NSError)
-    }
-    
-    func grantAnalyticsPermission() {
-        hasAcceptedAnalytics = true
-    }
-    
-    func denyAnalyticsPermission() {
-        hasAcceptedAnalytics = false
+
+    var _stubbedPermissionState: AnalyticsPermissionState = .accepted
+    var permissionState: AnalyticsPermissionState {
+        _stubbedPermissionState
     }
 }
